Create codebook in garden, and improve github and s3 steps
